+++ app/(userSection)/submissions/edit/[preprint]/page.tsx
@@ -17,7 +17,7 @@ const Page = async ({ params }: { params: { preprint: string } }) => {
     ),
   ])
   if (versionsRes.status !== 200 || preprintRes.status !== 200) {
-    redirect('/login?signOut=true')
+    redirect('/account?signOut=true')
   }
   const [versions, preprint] = await Promise.all([
     versionsRes.json(),
