+++ middleware.ts
@@ -2,14 +2,13 @@ import { withAuth } from 'next-auth/middleware'
 
 export const middleware = withAuth({
   pages: {
-    signIn: '/login',
+    signIn: '/account',
   },
 })
 
 export const config = {
   matcher: [
     '/submissions',
-    '/account',
     '/submit/overview',
     '/submit/info',
     '/submit/authors',
